+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @xnetjs/crypto
 
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/core@2.4.0
+
 ## 2.3.0
 
 ### Patch Changes