+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xnetjs/plugins
 
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies [[`1c7b9c9`](https://github.com/crs48/xNet/commit/1c7b9c9c3804fc0d4c80b032ae0ebc0163714c52)]:
+  - @xnetjs/data@2.4.0
+  - @xnetjs/abuse@2.4.0
+  - @xnetjs/core@2.4.0
+
 ## 2.3.0
 
 ### Patch Changes